requirements.txt
pypattyrn~=1.2
onnxruntime~=1.16.3
pyyaml~=6.0.1
torch~=2.1.0+cu121
timm~=0.9.12
tqdm~=4.65.2

ultralytics~=8.0.234
numpy~=1.26.3
opencv-python~=4.8.1.78
pillow~=10.2.0
tabulate~=0.9.0
loguru~=0.7.2
pycocotools~=2.0.7
psutil~=5.9.6
torchvision~=0.16.0+cu121
setuptools~=69.0.3
matplotlib~=3.8.2
pandas~=2.1.4
seaborn~=0.13.1
scipy~=1.11.4
Flask~=2.3.2
thop~=0.1.1.post2209072238
addict~=2.4.0
pytest~=7.4.4
shapely~=2.0.2
mmcv~=2.1.0
mmengine~=0.10.2
mmdet~=3.2.0
rich~=13.4.2
six~=1.16.0
terminaltables~=3.1.10
Jinja2~=3.1.2
safetensors~=0.4.0
filetype~=1.2.0
lxml
mmsegmentation